initial commit
This app is a prototype built for Southern California Edison to demonstrate a mobile hybrid app for iOS and Android using a Xamarin Forms Cross Platform .Net Standard, MVVM Prism Unity architecture.

It uses Azure app services for authentication and REST database storage.

It also uses xamarin forms maps to simulate electrical workers in the field using their mobile devices to monitor electrical outages.

The architecture also uses dependency injection to inject api services into each view model so that a mobile api does not have to be instantiated in every view model but rather injected as a service.

The main purpose of this prototype is to demonstrate loose coupling between the view, view model and api services.

Thanks.